Custom cycle processor support
